Diurnal centroid calculation

Calculate the diurnal centroid for sapflow variables

diurnal_centroid(variable)

Arguments

variable
A numeric vector containing the sapflow values for a day at a regular intervals. Missing values are allowed but not recommended

Value

A numeric value with the diurnal centroid value (0 to 24 h)

Details

The code for this function has been kindly provided by Jacob Nelson in python (see https://github.com/jnelson18/FluxnetTools/blob/master/FileS3.py) and has been translated to a tidy data phylosophy in R to be used inside a summarise statement.

Diurnal centroid algorithm

Given a continuous subdaily sapflow values at regular intervals \(V = {x_1, ..., x_n}\) to obtain the diurnal centroid each value is multiplied by its interval index and summed up and divided by the sum of the values for the day and finally the value is normalized to 24h:

$$ \sum {x_1 * 1, x_2 * 2, ... , x_n * n} / \sum {x_1, x_2, ... , x_n} * (24/n) $$

With even values for all the intervals (i.e. 100 for all), centroid converges to 12h at more than 1000 intervals per day. With only 48 (half hourly measurements) centroid converges to 12.25h and with 24 intervals (hourly measurements) centroid converges to 12.5h. So, using diurnal centroid value in half hourly datasets or above can have a considerable error associated.

Examples

# dplyr
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# check convergence to 12h:
diurnal_centroid(rep(1, 1000)) # 12.012 h

#> [1] 12.012

diurnal_centroid(rep(10000, 1000)) # 12.012 h, variable scale not affects calculation

#> [1] 12.012

# sapflow diurnal centroid
data('ARG_TRE', package = 'sapfluxnetr')

sfn_metrics(
  ARG_TRE,
  period = 'daily',
  .funs = funs(diurnal_centroid(.),
               data_coverage(., timestep, period_minutes)),
  solar = FALSE,
  interval = 'general'
)

#> [1] "Crunching data for ARG_TRE. In large datasets this could take a while"
#> [1] "General data for ARG_TRE"

#> $sapf
#> # A time tibble: 13 x 9
#> # Index: TIMESTAMP
#>    TIMESTAMP           ARG_TRE_Nan_Jt_… ARG_TRE_Nan_Jt_… ARG_TRE_Nan_Jt_…
#>    <dttm>                         <dbl>            <dbl>            <dbl>
#>  1 2009-11-18 00:00:00             14.1             14.3             13.7
#>  2 2009-11-19 00:00:00             14.4             14.0             13.2
#>  3 2009-11-20 00:00:00             14.0             13.4             12.3
#>  4 2009-11-21 00:00:00             13.6             13.4             12.4
#>  5 2009-11-22 00:00:00             15.1             14.5             14.1
#>  6 2009-11-23 00:00:00             16.0             15.9             15.1
#>  7 2009-11-24 00:00:00             15.5             15.3             14.9
#>  8 2009-11-25 00:00:00             13.5             13.6             13.0
#>  9 2009-11-26 00:00:00             12.4             12.0             12.6
#> 10 2009-11-27 00:00:00             15.8             16.0             16.2
#> 11 2009-11-28 00:00:00             12.6             12.4             12.4
#> 12 2009-11-29 00:00:00             14.9             14.5             14.8
#> 13 2009-11-30 00:00:00             16.3             16.4             16.1
#> # … with 5 more variables: ARG_TRE_Nan_Jt_4_diurnal_centroid <dbl>,
#> #   ARG_TRE_Nan_Jt_1_data_coverage <dbl>, ARG_TRE_Nan_Jt_2_data_coverage <dbl>,
#> #   ARG_TRE_Nan_Jt_3_data_coverage <dbl>, ARG_TRE_Nan_Jt_4_data_coverage <dbl>
#> 
#> $env
#> # A time tibble: 13 x 19
#> # Index: TIMESTAMP
#>    TIMESTAMP           ta_diurnal_cent… rh_diurnal_cent… vpd_diurnal_cen…
#>    <dttm>                         <dbl>            <dbl>            <dbl>
#>  1 2009-11-18 00:00:00             15.6             10.8             14.4
#>  2 2009-11-19 00:00:00             14.0             10.9             13.8
#>  3 2009-11-20 00:00:00             14.3             12.5             12.9
#>  4 2009-11-21 00:00:00             15.2             12.4             13.2
#>  5 2009-11-22 00:00:00             15.9             11.3             14.0
#>  6 2009-11-23 00:00:00             18.2             10.7             14.8
#>  7 2009-11-24 00:00:00             17.0             10.2             15.0
#>  8 2009-11-25 00:00:00             12.9             12.5             12.7
#>  9 2009-11-26 00:00:00             12.7             13.5             12.1
#> 10 2009-11-27 00:00:00             15.3             10.5             14.7
#> 11 2009-11-28 00:00:00             12.0             13.5             11.9
#> 12 2009-11-29 00:00:00             16.7             12.6             13.1
#> 13 2009-11-30 00:00:00             15.7             11.6             13.8
#> # … with 15 more variables: sw_in_diurnal_centroid <dbl>,
#> #   ws_diurnal_centroid <dbl>, precip_diurnal_centroid <dbl>,
#> #   swc_shallow_diurnal_centroid <dbl>, ppfd_in_diurnal_centroid <dbl>,
#> #   ext_rad_diurnal_centroid <dbl>, ta_data_coverage <dbl>,
#> #   rh_data_coverage <dbl>, vpd_data_coverage <dbl>, sw_in_data_coverage <dbl>,
#> #   ws_data_coverage <dbl>, precip_data_coverage <dbl>,
#> #   swc_shallow_data_coverage <dbl>, ppfd_in_data_coverage <dbl>,
#> #   ext_rad_data_coverage <dbl>
#>