Filter the sites by metadata variable values

filter_sites_by_md function takes logical expressions for the metadata variables (i.e. pl_sens_meth == 'HR'), and list the sites that met the criteria from thos supplied

filter_sites_by_md(sites, metadata, ..., .join = c("and", "or"))

Arguments

sites
character vector with the sites codes to filter, generally the result of sfn_sites_in_folder

metadata
metadata tbl object, usually the result of read_sfn_metadata

...
Logical expressions for the metadata variables, as in filter.

.join
Character indicating how to filter the sites, see details.

Value

A character vector with the sites fullfilling the premises

Details

.join argument indicates how sites must be filtered between metadata classes. 'and' indicates only sites meeting all conditions for all metadata classes are returned. 'or' indicates all sites meeting any condition between classes are returned. For two or more filtes of the same metadata class, they are combined as 'and'.

Examples

# simple, want to know which sites are using the Heat Ratio method to measure
# the sap flow.

# For this to work there must be a folder called 'Data' in the working
# directory containing the sites .RData files

# NOT RUN {
sites <- sfn_sites_in_folder('Data')
metadata <- read_sfn_metadata('Data')

filter_sites_by_md(pl_sens_meth == 'HR', sites = sites, metadata = metadata)

# Both, Heat Ratio and Heat Dissipation
filter_sites_by_md(
  pl_sens_meth %in% c('HR', 'HD'),
  sites = sites, metadata = metadata
)

# more complex, Heat Ratio method AND Mediterranean biome
filter_sites_by_md(
  pl_sens_meth == 'HR',
  si_biome == 'Mediterranean',
  sites = sites, metadata = metadata
  .join = 'and' # default
)

# join = 'or' returns sites that meet any condition
filter_sites_by_md(
  pl_sens_meth == 'HR',
  si_biome == 'Mediterranean',
  sites = sites, metadata = metadata
  .join = 'or'
)
# }