Build a tidy data frame from the metrics results nested list

Transform the nested list of metrics in a tidy tibble where each observation has its own row

metrics_tidyfier(metrics_res, metadata, interval = c("general",
  "predawn", "midday", "night", "daylight"))

Arguments

metrics_res
Nested list containing the metrics results as obtained from metrics

metadata
List containing the metadata nested list, as obtained from read_sfn_metadata

interval
Interval to return, it depends on the metrics_res and can be "gen" for the general metrics, "md" for midday metrics, "pd" for predawn metrics, "night" for night metrics or "day" for diurnal metrics.

Value

a tibble with the following columns:

TIMESTAMP: POSIXct vector with the date-time of the observation

si_code: Character vector with the site codes

pl_code: Character vector with the plant codes

sapflow_*: Variables containing the different metrics for the sapflow measurements (i.e. sapflow_mean, sapflow_q_95)

ta_*; rh_*; vpd_*; ...: Variables containing the different metrics for environmental variables (i.e. ta_mean, ta_q_95)

pl_*: plant metadata variables (i.e. pl_sapw_area, pl_sens_meth)

si_*: site metadata variables (i.e. si_biome, si_contact_firstname)

st_*: stand metadata variables (i.e. st_aspect, st_lai)

sp_*: species metadata variables (i.e. sp_basal_area_perc)

env_*: environmental metadata variables (i.e. env_timezone)

Examples

# data
multi_sfn <- sfn_data_multi(ARG_TRE, ARG_MAZ, AUS_CAN_ST2_MIX)
data('sfn_metadata_ex', package = 'sapfluxnetr')

# metrics
multi_metrics <- daily_metrics(multi_sfn)

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

#> [1] "Crunching data for ARG_TRE. In large datasets this could take a while"
#> [1] "General data for ARG_TRE"
#> [1] "Crunching data for ARG_MAZ. In large datasets this could take a while"
#> [1] "General data for ARG_MAZ"
#> [1] "Crunching data for AUS_CAN_ST2_MIX. In large datasets this could take a while"
#> [1] "General data for AUS_CAN_ST2_MIX"

# tidyfing
multi_tidy <- metrics_tidyfier(
  multi_metrics, sfn_metadata_ex, interval = 'general'
)
multi_tidy

#> # A tibble: 12,769 x 129
#>    TIMESTAMP           si_code pl_code sapflow_centroid sapflow_coverage
#>    <dttm>              <chr>   <chr>              <dbl>            <dbl>
#>  1 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#>  2 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#>  3 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#>  4 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#>  5 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#>  6 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#>  7 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#>  8 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#>  9 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#> 10 2006-06-20 00:00:00 AUS_CA… AUS_CA…              NaN                0
#> # … with 12,759 more rows, and 124 more variables: sapflow_mean <dbl>,
#> #   sapflow_q_95 <dbl>, sapflow_sd <dbl>, ta_mean <dbl>, rh_mean <dbl>,
#> #   vpd_mean <dbl>, sw_in_mean <dbl>, ws_mean <dbl>, precip_mean <dbl>,
#> #   swc_shallow_mean <dbl>, ppfd_in_mean <dbl>, ext_rad_mean <dbl>,
#> #   ta_sd <dbl>, rh_sd <dbl>, vpd_sd <dbl>, sw_in_sd <dbl>, ws_sd <dbl>,
#> #   precip_sd <dbl>, swc_shallow_sd <dbl>, ppfd_in_sd <dbl>, ext_rad_sd <dbl>,
#> #   ta_coverage <dbl>, rh_coverage <dbl>, vpd_coverage <dbl>,
#> #   sw_in_coverage <dbl>, ws_coverage <dbl>, precip_coverage <dbl>,
#> #   swc_shallow_coverage <dbl>, ppfd_in_coverage <dbl>, ext_rad_coverage <dbl>,
#> #   ta_q_95 <dbl>, rh_q_95 <dbl>, vpd_q_95 <dbl>, sw_in_q_95 <dbl>,
#> #   ws_q_95 <dbl>, precip_q_95 <dbl>, swc_shallow_q_95 <dbl>,
#> #   ppfd_in_q_95 <dbl>, ext_rad_q_95 <dbl>, precip_accumulated <dbl>,
#> #   pl_age <dbl>, pl_azimut_int <chr>, pl_bark_thick <dbl>, pl_dbh <dbl>,
#> #   pl_height <dbl>, pl_leaf_area <dbl>, pl_name <chr>, pl_radial_int <chr>,
#> #   pl_remarks <chr>, pl_sap_units <chr>, pl_sapw_area <dbl>,
#> #   pl_sapw_depth <dbl>, pl_sens_calib <lgl>, pl_sens_cor_grad <chr>,
#> #   pl_sens_cor_zero <chr>, pl_sens_hgt <dbl>, pl_sens_length <int>,
#> #   pl_sens_man <chr>, pl_sens_meth <chr>, pl_sens_timestep <int>,
#> #   pl_social <chr>, pl_species <chr>, pl_treatment <chr>,
#> #   pl_sap_units_orig <chr>, si_addcontr_email <chr>,
#> #   si_addcontr_firstname <chr>, si_addcontr_institution <chr>,
#> #   si_addcontr_lastname <chr>, si_contact_email <chr>,
#> #   si_contact_firstname <chr>, si_contact_institution <chr>,
#> #   si_contact_lastname <chr>, si_country <chr>, si_dendro_network <lgl>,
#> #   si_dist_mgmt <chr>, si_elev <int>, si_flux_network <lgl>, si_igbp <chr>,
#> #   si_lat <dbl>, si_long <dbl>, si_name <chr>, si_paper <chr>,
#> #   si_remarks <chr>, is_inside_country <lgl>, si_mat <dbl>, si_map <dbl>,
#> #   si_biome <chr>, st_age <dbl>, st_aspect <chr>, st_basal_area <dbl>,
#> #   st_clay_perc <lgl>, st_density <int>, st_growth_condition <chr>,
#> #   st_height <dbl>, st_lai <dbl>, st_name <chr>, st_remarks <lgl>,
#> #   st_sand_perc <lgl>, st_silt_perc <lgl>, st_soil_depth <int>, …

# A really easier way of doing the same
multi_tidy_easy <- daily_metrics(multi_sfn, tidy = TRUE, metadata = sfn_metadata_ex)

#> [1] "Crunching data for ARG_TRE. In large datasets this could take a while"
#> [1] "General data for ARG_TRE"
#> [1] "Crunching data for ARG_MAZ. In large datasets this could take a while"
#> [1] "General data for ARG_MAZ"
#> [1] "Crunching data for AUS_CAN_ST2_MIX. In large datasets this could take a while"
#> [1] "General data for AUS_CAN_ST2_MIX"