S4 class for sapfluxnet site data

Main class for storing sapfluxnet project site data and metadata

Details

This class allows to store all the data and metadata for a sapfluxnet site in one single object, to easily work with it. See vignette('sfn-data-classes', package = 'sapfluxnetr') for more info.

Slots

sapf_data

A data frame with the sapf data

env_data

A data frame with the env data

sapf_flags

A data frame with the same dimensions of sapf_data with the flag info for each tree/TIMESTAMP combination

env_flags

A data frame with the same dimensions of env_data with the flag info for each env_var/TIMESTAMP combination

si_code

A character vector of length one indicating the site code

timestamp

A POSIXct vector of length nrow(sapf_data) with the timestamp

solar_timestamp

A POSIXct vector of length nrow(sapf_data) with the solar timestamp

site_md

A data frame containing the site metadata

stand_md

A data frame containing the stand metadata

species_md

A data frame containing the species metadata

plant_md

A data frame containing the plant metadata

env_md

A data frame containing the env metadata