sfn_metadata cache file for example data (ARG_MAZ, ARG_TRE and AUS_CAN_ST2_MIX)

Example metadata cache file content for package usage demonstration

sfn_metadata_ex

Format

A list with five elements, each of one a metadata type.

Examples

data('sfn_metadata_ex', package = 'sapfluxnetr')
sfn_metadata_ex

#> $site_md
#> # A tibble: 3 x 24
#>   si_addcontr_ema… si_addcontr_fir… si_addcontr_ins… si_addcontr_las… si_code
#>   <chr>            <chr>            <chr>            <chr>            <chr>  
#> 1 NA               Pablo            Instituto Nacio… Peri             ARG_MAZ
#> 2 NA               Pablo            Instituto Nacio… Peri             ARG_TRE
#> 3 NA               NA               NA               NA               AUS_CA…
#> # … with 19 more variables: si_contact_email <chr>, si_contact_firstname <chr>,
#> #   si_contact_institution <chr>, si_contact_lastname <chr>, si_country <chr>,
#> #   si_dendro_network <lgl>, si_dist_mgmt <chr>, si_elev <int>,
#> #   si_flux_network <lgl>, si_igbp <chr>, si_lat <dbl>, si_long <dbl>,
#> #   si_name <chr>, si_paper <chr>, si_remarks <chr>, is_inside_country <lgl>,
#> #   si_mat <dbl>, si_map <dbl>, si_biome <chr>
#> 
#> $stand_md
#> # A tibble: 3 x 18
#>   st_age st_aspect st_basal_area st_clay_perc st_density st_growth_condi…
#>    <dbl> <chr>             <dbl> <lgl>             <int> <chr>           
#> 1  180   E                  59.1 NA                  518 Naturally regen…
#> 2  140   Flat               19.4 NA                  512 Naturally regen…
#> 3   14.5 Flat               24.4 NA                 1420 Plantation, unm…
#> # … with 12 more variables: st_height <dbl>, st_lai <dbl>, st_name <chr>,
#> #   st_remarks <lgl>, st_sand_perc <lgl>, st_silt_perc <lgl>,
#> #   st_soil_depth <int>, st_soil_texture <chr>, st_terrain <chr>,
#> #   st_treatment <chr>, si_code <chr>, st_USDA_soil_texture <chr>
#> 
#> $species_md
#> # A tibble: 4 x 5
#>   sp_basal_area_perc sp_leaf_habit  sp_name             sp_ntrees si_code       
#>                <dbl> <chr>          <chr>                   <int> <chr>         
#> 1              100   cold deciduous Nothofagus pumilio          5 ARG_MAZ       
#> 2              100   cold deciduous Nothofagus antarct…         4 ARG_TRE       
#> 3               55.0 evergreen      Eucalyptus globulus        18 AUS_CAN_ST2_M…
#> 4               45.0 evergreen      Acacia mearnsii            16 AUS_CAN_ST2_M…
#> 
#> $plant_md
#> # A tibble: 43 x 26
#>    pl_age pl_azimut_int pl_bark_thick pl_code pl_dbh pl_height pl_leaf_area
#>     <dbl> <chr>                 <dbl> <chr>    <dbl>     <dbl>        <dbl>
#>  1  180   Corrected, m…          14.7 ARG_MA…   41.1      NA          109. 
#>  2  180   No azimuthal…          13   ARG_MA…   33.2      NA           58.4
#>  3  180   No azimuthal…           6   ARG_MA…   23.2      NA           35.0
#>  4  180   No azimuthal…          21   ARG_MA…   55.6      NA          174. 
#>  5  180   No azimuthal…          10   ARG_MA…   38        NA           88  
#>  6  140   Corrected, m…          23.5 ARG_TR…   23.6      NA           29.7
#>  7  140   No azimuthal…          22.5 ARG_TR…   20.2      NA           13.0
#>  8  140   No azimuthal…          22   ARG_TR…   23.8      NA           29.8
#>  9  140   No azimuthal…          23.5 ARG_TR…   19.5      NA           15.1
#> 10   14.5 No azimuthal…          NA   AUS_CA…   13.1      14.3         NA  
#> # … with 33 more rows, and 19 more variables: pl_name <chr>,
#> #   pl_radial_int <chr>, pl_remarks <chr>, pl_sap_units <chr>,
#> #   pl_sapw_area <dbl>, pl_sapw_depth <dbl>, pl_sens_calib <lgl>,
#> #   pl_sens_cor_grad <chr>, pl_sens_cor_zero <chr>, pl_sens_hgt <dbl>,
#> #   pl_sens_length <int>, pl_sens_man <chr>, pl_sens_meth <chr>,
#> #   pl_sens_timestep <int>, pl_social <chr>, pl_species <chr>,
#> #   pl_treatment <chr>, si_code <chr>, pl_sap_units_orig <chr>
#> 
#> $env_md
#> # A tibble: 3 x 17
#>   env_leafarea_se… env_netrad env_plant_watpot env_ppfd_in env_precip
#>   <lgl>            <chr>      <chr>            <chr>       <chr>     
#> 1 NA               Not provi… leaf: midday     Not provid… Clearing  
#> 2 NA               Not provi… leaf: midday     Not provid… Clearing  
#> 3 NA               Not provi… NA               Not provid… Clearing  
#> # … with 12 more variables: env_remarks <chr>, env_rh <chr>,
#> #   env_swc_deep_depth <lgl>, env_swc_shallow_depth <int>, env_sw_in <chr>,
#> #   env_ta <chr>, env_time_daylight <lgl>, env_timestep <int>,
#> #   env_time_zone <chr>, env_vpd <chr>, env_ws <chr>, si_code <chr>
#>