Metrics summary function

Generate metrics from a site/s data for the period indicated

sfn_metrics(sfn_data, period, .funs, solar, interval = c("general",
  "predawn", "midday", "night", "daylight"), int_start = NULL,
  int_end = NULL, ...)

Arguments

sfn_data
sfn_data or sfn_data_multi object to obtain the metrics from

period
Time period to aggregate data by. See period section for an explanation about the periods ('daily', 'monthly', 'yearly', ...)

.funs
List of function calls to summarise the data by, usually the result of calling funs

solar
Logical indicating if the solarTIMESTAMP must be used instead of the site local TIMESTAMP. Default to TRUE (use solarTIMESTAMP).

interval
Character vector indicating if the metrics must be filtered by an special hour interval. See Interval section in details.

int_start
Integer value indicating the starting hour of the special interval in 24h format. See Interval section in details.

int_end
Integer value indicating the ending hour of the special interval in 24h format. See Interval section in details.

...
optional arguments to pass to methods used (i.e. tibbletime::collapse_index or summarise funs extra arguments)

Value

For sfn_data objects, a list of tbl_time objects with the following structure:

$sapf: metrics for the sapflow data

$env: metrics for the environmental data

For sfn_data_multi objects, a list of lists of tbl_time objects with the metrics for each site:

$SITE_CODE

$sapf: metrics for the sapflow data

$env: metrics for the environmental data

$NEXT_SITE_CODE...

Period

period argument is piped to tibbletime::collapse_index function with side = 'start', clean = TRUE options. See collapse_index for a detailed explanation, but in short:

frequency period format: "1 day", "7 day", "1 month", "1 year"

shorthand format: "hourly", "daily", "monthly", "yearly"

custom format: a vector of dates to use as custom and more flexible boundaries

Also, you can override the default behaviour of sfn_metrics, providing side = 'end' or clean = FALSE.

.funs

.funs argument uses the same method as the .funs argument in the summarise_all function of dplyr package. Basically it accepts a list of function calls generated by funs(), or a character vector of function names, or simply a function. If you want to pass on a custom function you can specify it here. See details in summarise_by_period for more complex summarising functions declaration.

Interval

Previously to the metrics summary, data can be filtered by an special interval (predawn for example). This filtering can be specified with the interval argument this:

general (default). No special interval is used, and metrics are performed with all the data.

predawn. Data is filtered for predawn interval. In this case int_start and int_end must be specified as 24h value

midday. Data is filtered for midday interval. In this case int_start and int_end must be specified as 24h value

night. Data is filtered for night interval. In this case int_start and int_end must be specified as 24h value

daylight. Data is filtered for daylight interval. In this case int_start and int_end must be specified as 24h value

See also

Examples

library(dplyr)

### general metrics
## sfn_data
data('ARG_TRE', package = 'sapfluxnetr')
ARG_TRE_metrics <- sfn_metrics(
  ARG_TRE,
  period = '7 days',
  .funs = funs(mean(., na.rm = TRUE), sd(., na.rm = TRUE), n()),
  solar = FALSE,
  interval = 'general'
)

#> [1] "Crunching data for ARG_TRE. In large datasets this could take a while"
#> [1] "General data for ARG_TRE"

str(ARG_TRE_metrics)

#> List of 2
#>  $ sapf:Classes ‘tbl_time’, ‘tbl_df’, ‘tbl’ and 'data.frame':	3 obs. of  13 variables:
#>   ..$ TIMESTAMP            : POSIXct[1:3], format: "2009-11-15" "2009-11-22" ...
#>   ..$ ARG_TRE_Nan_Jt_1_mean: num [1:3] 433 522 374
#>   ..$ ARG_TRE_Nan_Jt_2_mean: num [1:3] 300 364 286
#>   ..$ ARG_TRE_Nan_Jt_3_mean: num [1:3] 346 460 365
#>   ..$ ARG_TRE_Nan_Jt_4_mean: num [1:3] 428 620 564
#>   ..$ ARG_TRE_Nan_Jt_1_sd  : num [1:3] 211 272 245
#>   ..$ ARG_TRE_Nan_Jt_2_sd  : num [1:3] 156 205 209
#>   ..$ ARG_TRE_Nan_Jt_3_sd  : num [1:3] 136 256 238
#>   ..$ ARG_TRE_Nan_Jt_4_sd  : num [1:3] 206 347 456
#>   ..$ ARG_TRE_Nan_Jt_1_n   : int [1:3] 96 168 48
#>   ..$ ARG_TRE_Nan_Jt_2_n   : int [1:3] 96 168 48
#>   ..$ ARG_TRE_Nan_Jt_3_n   : int [1:3] 96 168 48
#>   ..$ ARG_TRE_Nan_Jt_4_n   : int [1:3] 96 168 48
#>   ..- attr(*, "index_quo")= language ~TIMESTAMP
#>   .. ..- attr(*, ".Environment")=<environment: 0x55ce10dea978> 
#>   ..- attr(*, "index_time_zone")= chr "Etc/GMT+3"
#>  $ env :Classes ‘tbl_time’, ‘tbl_df’, ‘tbl’ and 'data.frame':	3 obs. of  28 variables:
#>   ..$ TIMESTAMP       : POSIXct[1:3], format: "2009-11-15" "2009-11-22" ...
#>   ..$ ta_mean         : num [1:3] 4.91 6.15 2.48
#>   ..$ rh_mean         : num [1:3] 34.9 35.4 41.1
#>   ..$ vpd_mean        : num [1:3] 0.605 0.656 0.449
#>   ..$ sw_in_mean      : num [1:3] 286 327 250
#>   ..$ ws_mean         : num [1:3] 15.6 24.6 22.4
#>   ..$ precip_mean     : num [1:3] 0.00625 0.19226 0.11667
#>   ..$ swc_shallow_mean: num [1:3] 0.366 0.348 0.373
#>   ..$ ppfd_in_mean    : num [1:3] 604 691 528
#>   ..$ ext_rad_mean    : num [1:3] 473 486 496
#>   ..$ ta_sd           : num [1:3] 3.75 4.01 2.82
#>   ..$ rh_sd           : num [1:3] 10.8 12.3 13.1
#>   ..$ vpd_sd          : num [1:3] 0.265 0.274 0.17
#>   ..$ sw_in_sd        : num [1:3] 358 394 295
#>   ..$ ws_sd           : num [1:3] 6.57 14.65 12.97
#>   ..$ precip_sd       : num [1:3] 0.0431 0.9483 0.3581
#>   ..$ swc_shallow_sd  : num [1:3] 0.0157 0.031 0.0267
#>   ..$ ppfd_in_sd      : num [1:3] 756 832 623
#>   ..$ ext_rad_sd      : num [1:3] 464 469 476
#>   ..$ ta_n            : int [1:3] 96 168 48
#>   ..$ rh_n            : int [1:3] 96 168 48
#>   ..$ vpd_n           : int [1:3] 96 168 48
#>   ..$ sw_in_n         : int [1:3] 96 168 48
#>   ..$ ws_n            : int [1:3] 96 168 48
#>   ..$ precip_n        : int [1:3] 96 168 48
#>   ..$ swc_shallow_n   : int [1:3] 96 168 48
#>   ..$ ppfd_in_n       : int [1:3] 96 168 48
#>   ..$ ext_rad_n       : int [1:3] 96 168 48
#>   ..- attr(*, "index_quo")= language ~TIMESTAMP
#>   .. ..- attr(*, ".Environment")=<environment: 0x55ce1053c278> 
#>   ..- attr(*, "index_time_zone")= chr "Etc/GMT+3"

ARG_TRE_metrics[['sapf']]

#> # A time tibble: 3 x 13
#> # Index: TIMESTAMP
#>   TIMESTAMP           ARG_TRE_Nan_Jt_… ARG_TRE_Nan_Jt_… ARG_TRE_Nan_Jt_…
#>   <dttm>                         <dbl>            <dbl>            <dbl>
#> 1 2009-11-15 00:00:00             433.             300.             346.
#> 2 2009-11-22 00:00:00             522.             364.             460.
#> 3 2009-11-29 00:00:00             374.             286.             365.
#> # … with 9 more variables: ARG_TRE_Nan_Jt_4_mean <dbl>,
#> #   ARG_TRE_Nan_Jt_1_sd <dbl>, ARG_TRE_Nan_Jt_2_sd <dbl>,
#> #   ARG_TRE_Nan_Jt_3_sd <dbl>, ARG_TRE_Nan_Jt_4_sd <dbl>,
#> #   ARG_TRE_Nan_Jt_1_n <int>, ARG_TRE_Nan_Jt_2_n <int>,
#> #   ARG_TRE_Nan_Jt_3_n <int>, ARG_TRE_Nan_Jt_4_n <int>

ARG_TRE_metrics[['env']]

#> # A time tibble: 3 x 28
#> # Index: TIMESTAMP
#>   TIMESTAMP           ta_mean rh_mean vpd_mean sw_in_mean ws_mean precip_mean
#>   <dttm>                <dbl>   <dbl>    <dbl>      <dbl>   <dbl>       <dbl>
#> 1 2009-11-15 00:00:00    4.91    34.9    0.605       286.    15.6     0.00625
#> 2 2009-11-22 00:00:00    6.15    35.4    0.656       327.    24.6     0.192  
#> 3 2009-11-29 00:00:00    2.48    41.1    0.449       250.    22.4     0.117  
#> # … with 21 more variables: swc_shallow_mean <dbl>, ppfd_in_mean <dbl>,
#> #   ext_rad_mean <dbl>, ta_sd <dbl>, rh_sd <dbl>, vpd_sd <dbl>, sw_in_sd <dbl>,
#> #   ws_sd <dbl>, precip_sd <dbl>, swc_shallow_sd <dbl>, ppfd_in_sd <dbl>,
#> #   ext_rad_sd <dbl>, ta_n <int>, rh_n <int>, vpd_n <int>, sw_in_n <int>,
#> #   ws_n <int>, precip_n <int>, swc_shallow_n <int>, ppfd_in_n <int>,
#> #   ext_rad_n <int>

## sfn_data_multi

# NOT RUN {
data('ARG_MAZ', package = 'sapfluxnetr')
data('AUS_CAN_ST2_MIX', package = 'sapfluxnetr')
multi_sfn <- sfn_data_multi(ARG_TRE, ARG_MAZ, AUS_CAN_ST2_MIX)

multi_metrics <- sfn_metrics(
  multi_sfn,
  period = '7 days',
  .funs = funs(mean(., na.rm = TRUE), sd(., na.rm = TRUE), n()),
  solar = FALSE,
  interval = 'general'
)

str(multi_metrics)

multi_metrics[['ARG_TRE']][['sapf']]
# }

### midday metrics
ARG_TRE_midday <- sfn_metrics(
  ARG_TRE,
  period = 'daily',
  .funs = funs(mean(., na.rm = TRUE), sd(., na.rm = TRUE), n()),
  solar = TRUE,
  interval = 'midday', int_start = 11, int_end = 13
)

#> [1] "Crunching data for ARG_TRE. In large datasets this could take a while"
#> [1] "midday data for ARG_TRE"

str(ARG_TRE_midday)

#> List of 2
#>  $ sapf:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	13 obs. of  13 variables:
#>   ..$ TIMESTAMP_md            : POSIXct[1:13], format: "2009-11-18" "2009-11-19" ...
#>   ..$ ARG_TRE_Nan_Jt_1_mean_md: num [1:13] 685 879 438 631 783 ...
#>   ..$ ARG_TRE_Nan_Jt_2_mean_md: num [1:13] 665 594 272 379 535 ...
#>   ..$ ARG_TRE_Nan_Jt_3_mean_md: num [1:13] 614 626 258 533 680 ...
#>   ..$ ARG_TRE_Nan_Jt_4_mean_md: num [1:13] 719 664 474 619 875 ...
#>   ..$ ARG_TRE_Nan_Jt_1_sd_md  : num [1:13] 23.8 192.6 115.7 46.8 40.9 ...
#>   ..$ ARG_TRE_Nan_Jt_2_sd_md  : num [1:13] 70.1 67.3 54.2 25.1 42.4 ...
#>   ..$ ARG_TRE_Nan_Jt_3_sd_md  : num [1:13] 40.14 9.82 63.06 7.69 193.53 ...
#>   ..$ ARG_TRE_Nan_Jt_4_sd_md  : num [1:13] 102.15 2.9 59.78 19.73 2.32 ...
#>   ..$ ARG_TRE_Nan_Jt_1_n_md   : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ ARG_TRE_Nan_Jt_2_n_md   : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ ARG_TRE_Nan_Jt_3_n_md   : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ ARG_TRE_Nan_Jt_4_n_md   : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..- attr(*, "index_quo")= language ~TIMESTAMP
#>   .. ..- attr(*, ".Environment")=<environment: 0x55ce1afca048> 
#>   ..- attr(*, "index_time_zone")= chr "UTC"
#>  $ env :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	13 obs. of  28 variables:
#>   ..$ TIMESTAMP_md       : POSIXct[1:13], format: "2009-11-18" "2009-11-19" ...
#>   ..$ ta_mean_md         : num [1:13] 10.5 12.5 5.38 6.7 8.15 ...
#>   ..$ rh_mean_md         : num [1:13] 24.2 15.9 37.2 31.2 22.6 ...
#>   ..$ vpd_mean_md        : num [1:13] 0.963 1.22 0.563 0.676 0.84 ...
#>   ..$ sw_in_mean_md      : num [1:13] 979 1126 218 499 1077 ...
#>   ..$ ws_mean_md         : num [1:13] 21.4 20.3 19.3 27 36.1 ...
#>   ..$ precip_mean_md     : num [1:13] 0 0 0 0 0 0 0 0 3.6 0 ...
#>   ..$ swc_shallow_mean_md: num [1:13] 0.386 0.37 0.356 0.348 0.343 ...
#>   ..$ ppfd_in_mean_md    : num [1:13] 2070 2381 462 1054 2277 ...
#>   ..$ ext_rad_mean_md    : num [1:13] 1179 1183 1187 1190 1193 ...
#>   ..$ ta_sd_md           : num [1:13] 0 0.566 1.025 0.283 1.202 ...
#>   ..$ rh_sd_md           : num [1:13] 0.747 0.179 2.511 0.209 1.285 ...
#>   ..$ vpd_sd_md          : num [1:13] 0.00949 0.0479 0.01766 0.01521 0.05468 ...
#>   ..$ sw_in_sd_md        : num [1:13] 427 330 107 255 102 ...
#>   ..$ ws_sd_md           : num [1:13] 1.58 6.34 2.72 0 1.3 ...
#>   ..$ precip_sd_md       : num [1:13] 0 0 0 0 0 ...
#>   ..$ swc_shallow_sd_md  : num [1:13] 0.001011 0.001513 0.001011 0 0.000396 ...
#>   ..$ ppfd_in_sd_md      : num [1:13] 902 697 226 540 215 ...
#>   ..$ ext_rad_sd_md      : num [1:13] 6.32 5.98 5.64 5.27 4.9 ...
#>   ..$ ta_n_md            : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ rh_n_md            : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ vpd_n_md           : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ sw_in_n_md         : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ ws_n_md            : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ precip_n_md        : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ swc_shallow_n_md   : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ ppfd_in_n_md       : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ ext_rad_n_md       : int [1:13] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..- attr(*, "index_quo")= language ~TIMESTAMP
#>   .. ..- attr(*, ".Environment")=<environment: 0x55ce191cd238> 
#>   ..- attr(*, "index_time_zone")= chr "UTC"

ARG_TRE_midday[['sapf']]

#> # A tibble: 13 x 13
#>    TIMESTAMP_md        ARG_TRE_Nan_Jt_… ARG_TRE_Nan_Jt_… ARG_TRE_Nan_Jt_…
#>    <dttm>                         <dbl>            <dbl>            <dbl>
#>  1 2009-11-18 00:00:00             685.             665.             614.
#>  2 2009-11-19 00:00:00             879.             594.             626.
#>  3 2009-11-20 00:00:00             438.             272.             258.
#>  4 2009-11-21 00:00:00             631.             379.             533.
#>  5 2009-11-22 00:00:00             783.             535.             680.
#>  6 2009-11-23 00:00:00             841.             478.             618.
#>  7 2009-11-24 00:00:00             951.             636.             789.
#>  8 2009-11-25 00:00:00             907.             602.             789.
#>  9 2009-11-26 00:00:00             861.             697.             925.
#> 10 2009-11-27 00:00:00             806.             594.             706.
#> 11 2009-11-28 00:00:00             837.             730.             925.
#> 12 2009-11-29 00:00:00             638.             605.             666.
#> 13 2009-11-30 00:00:00             548.             371.             444.
#> # … with 9 more variables: ARG_TRE_Nan_Jt_4_mean_md <dbl>,
#> #   ARG_TRE_Nan_Jt_1_sd_md <dbl>, ARG_TRE_Nan_Jt_2_sd_md <dbl>,
#> #   ARG_TRE_Nan_Jt_3_sd_md <dbl>, ARG_TRE_Nan_Jt_4_sd_md <dbl>,
#> #   ARG_TRE_Nan_Jt_1_n_md <int>, ARG_TRE_Nan_Jt_2_n_md <int>,
#> #   ARG_TRE_Nan_Jt_3_n_md <int>, ARG_TRE_Nan_Jt_4_n_md <int>