List all variables that can be used to filter sites

sfn_vars_to_filter() returns a list with the variables for each kind of metadata that can be used to select and filter sites

sfn_vars_to_filter()

Value

A list with five elements, site_md, stand_md, species_md, plant_md and env_md

Examples

# all variables
sfn_vars_to_filter()

#> $site_md
#>  [1] "si_name"                 "si_country"             
#>  [3] "si_contact_firstname"    "si_contact_lastname"    
#>  [5] "si_contact_email"        "si_contact_institution" 
#>  [7] "si_addcontr_firstname"   "si_addcontr_lastname"   
#>  [9] "si_addcontr_email"       "si_addcontr_institution"
#> [11] "si_lat"                  "si_long"                
#> [13] "si_elev"                 "si_paper"               
#> [15] "si_dist_mgmt"            "si_igbp"                
#> [17] "si_flux_network"         "si_dendro_network"      
#> [19] "si_remarks"              "si_code"                
#> [21] "si_mat"                  "si_map"                 
#> [23] "si_biome"               
#> 
#> $stand_md
#>  [1] "st_name"              "st_growth_condition"  "st_treatment"        
#>  [4] "st_age"               "st_height"            "st_density"          
#>  [7] "st_basal_area"        "st_lai"               "st_aspect"           
#> [10] "st_terrain"           "st_soil_depth"        "st_soil_texture"     
#> [13] "st_sand_perc"         "st_silt_perc"         "st_clay_perc"        
#> [16] "st_remarks"           "st_USDA_soil_texture"
#> 
#> $species_md
#> [1] "sp_name"            "sp_ntrees"          "sp_leaf_habit"     
#> [4] "sp_basal_area_perc"
#> 
#> $plant_md
#>  [1] "pl_name"           "pl_species"        "pl_treatment"     
#>  [4] "pl_dbh"            "pl_height"         "pl_age"           
#>  [7] "pl_social"         "pl_sapw_area"      "pl_sapw_depth"    
#> [10] "pl_bark_thick"     "pl_leaf_area"      "pl_sens_meth"     
#> [13] "pl_sens_man"       "pl_sens_cor_grad"  "pl_sens_cor_zero" 
#> [16] "pl_sens_calib"     "pl_sap_units"      "pl_sap_units_orig"
#> [19] "pl_sens_length"    "pl_sens_hgt"       "pl_sens_timestep" 
#> [22] "pl_radial_int"     "pl_azimut_int"     "pl_remarks"       
#> [25] "pl_code"          
#> 
#> $env_md
#>  [1] "env_time_zone"         "env_time_daylight"     "env_timestep"         
#>  [4] "env_ta"                "env_rh"                "env_vpd"              
#>  [7] "env_sw_in"             "env_ppfd_in"           "env_netrad"           
#> [10] "env_ws"                "env_precip"            "env_swc_shallow_depth"
#> [13] "env_swc_deep_depth"    "env_plant_watpot"      "env_leafarea_seasonal"
#> [16] "env_remarks"          
#> 

# by some metadata
sfn_vars_to_filter()$site_md

#>  [1] "si_name"                 "si_country"             
#>  [3] "si_contact_firstname"    "si_contact_lastname"    
#>  [5] "si_contact_email"        "si_contact_institution" 
#>  [7] "si_addcontr_firstname"   "si_addcontr_lastname"   
#>  [9] "si_addcontr_email"       "si_addcontr_institution"
#> [11] "si_lat"                  "si_long"                
#> [13] "si_elev"                 "si_paper"               
#> [15] "si_dist_mgmt"            "si_igbp"                
#> [17] "si_flux_network"         "si_dendro_network"      
#> [19] "si_remarks"              "si_code"                
#> [21] "si_mat"                  "si_map"                 
#> [23] "si_biome"