Summaries by period

This function collapse the TIMESTAMP to the desired period (day, month...) by setting the same value to all timestamps within the period. This modified TIMESTAMP is used to group by and summarise the data.

summarise_by_period(data, period, .funs, ...)

Arguments

data
sapflow or environmental data as obtained by get_sapf_data and get_env_data. Must have a column named TIMESTAMP

period
tibbletime::collapse_index period

.funs
dplyr::summarise_all funs

...
optional arguments for link{summarise_by_period}

Value

A `tbl_time` object with the metrics results. The names of the columns indicate the original variable (tree or environmental variable) and the metric calculated (i.e. `vpd_mean`)

Details

This function uses internally collapse_index and summarise_all and arguments to control these functions can be passed as `...`. Arguments for each function are spliced and applied when needed. Be advised that all arguments passed to the summarise_all function will be applied to all the summarising functions used, so it will fail if any of that functions does not accept that argument. To complex function-argument relationships, indicate each summary function call within the .funs argument with funs:

# This will fail beacuse na.rm argument will be also passed to the n function,
# which does not accept any argument:
summarise_by_period(
  data = get_sapf_data(ARG_TRE),
  period = '7 days',
  .funs = funs(mean, sd, n),
  na.rm = TRUE
)
    # to solve this is better to use the .funs argument:
summarise_by_period(
  data = get_sapf_data(ARG_TRE),
  period = '7 days',
  .funs = funs(mean(., na.rm = TRUE), sd(., na.rm = TRUE), n())
)

TIMESTAMP_coll

Previously to the collapsing step, a temporal variable called TIMESTAMP_coll is created to be able to catch the real timestamp when some events happens, for example to use the min_time function. If your custom summarise function needs to get the time at which some event happens, use TIMESTAMP_coll instead of TIMESTAMP for that:

    min_time <- function(x, time) {
      time[which.min(x)]
    }
    summarise_by_period(
      data = get_sapf_data(ARG_TRE),
      period = 'daily',
      .funs = funs(min_time, time = TIMESTAMP_coll) # Not TIMESTAMP
    )
  

Examples

library(dplyr)

# data
data('ARG_TRE', package = 'sapfluxnetr')

# simple summary
summarise_by_period(
  data = get_sapf_data(ARG_TRE),
  period = '7 days',
  .funs = funs(mean(., na.rm = TRUE), sd(., na.rm = TRUE), n())
)

#> # A time tibble: 3 x 13
#> # Index: TIMESTAMP
#>   TIMESTAMP           ARG_TRE_Nan_Jt_… ARG_TRE_Nan_Jt_… ARG_TRE_Nan_Jt_…
#>   <dttm>                         <dbl>            <dbl>            <dbl>
#> 1 2009-11-15 00:00:00             433.             300.             346.
#> 2 2009-11-22 00:00:00             522.             364.             460.
#> 3 2009-11-29 00:00:00             374.             286.             365.
#> # … with 9 more variables: ARG_TRE_Nan_Jt_4_mean <dbl>,
#> #   ARG_TRE_Nan_Jt_1_sd <dbl>, ARG_TRE_Nan_Jt_2_sd <dbl>,
#> #   ARG_TRE_Nan_Jt_3_sd <dbl>, ARG_TRE_Nan_Jt_4_sd <dbl>,
#> #   ARG_TRE_Nan_Jt_1_n <int>, ARG_TRE_Nan_Jt_2_n <int>,
#> #   ARG_TRE_Nan_Jt_3_n <int>, ARG_TRE_Nan_Jt_4_n <int>