time at maximum/minimum

helpers for sfn_metrics

max_time(x, time)

min_time(x, time)

Arguments

x
a numeric vector, usually a variable in the sapflow or environmental data.

time
a POSIXct or character vector with the TIMESTAMP values

Value

a single value (POSIXct) with the TIMESTAMP value.

Details

These helper functions return the TIMESTAMP value at which the maximum value for other variable occurs. It is designed to be used inside a dplyr summarise statement.

Functions

max_time: helper for sfn_metrics

min_time: helper for sfn_metrics

Examples

library(dplyr)
storms %>%
  group_by(year) %>%
  summarise(wind_max = max(wind),
            hour_at_max = sapfluxnetr:::max_time(wind, time = hour),
            wind_min = min(wind),
            hour_at_min = sapfluxnetr:::min_time(wind, time = hour))

#> # A tibble: 41 x 5
#>     year wind_max hour_at_max wind_min hour_at_min
#>    <dbl>    <dbl>       <dbl>    <dbl>       <dbl>
#>  1  1975      100           0       20           6
#>  2  1976      105           0       20          12
#>  3  1977      150           6       20          12
#>  4  1978       80           0       20          12
#>  5  1979      150          18       15          12
#>  6  1980       90          18       20           0
#>  7  1981      115           0       20          12
#>  8  1982      115           0       20          12
#>  9  1983      100           6       20           0
#> 10  1984      115           0       15          18
#> # … with 31 more rows